+++ linux_kernel/network.h
@@ -11,5 +11,5 @@ extern unsigned int network_callback(const struct nf_hook_ops *ops,
                                      const struct net_device *in,
                                      const struct net_device *out,
                                      int (*okfn)(struct sk_buff *));
-extern void block_ip_address(size_t ip_address, size_t time_sec);
-extern bool check_is_blacklist_ip(size_t ip_address);
+extern void block_ip_address(u32 ip_address, size_t time_sec);
+extern bool check_is_blacklist_ip(u32 ip_address);
